+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 *   Bug Fixes
     *   Fix global auto download setting was incorrectly overriding the podcast auto download setting
         ([#3342](https://github.com/Automattic/pocket-casts-android/pull/3342))
+    *   Fix sleep timer was not stopping as expected
+        ([#3377](https://github.com/Automattic/pocket-casts-android/pull/3377))
 
 7.79
 -----